pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chesshacks-engine"
version = "0.1.0"
description = "Chess engine with neural network evaluation for ChessHacks2025"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "ChessHacks2025 Team"}
]

dependencies = [
    "numpy>=1.20.0",
    "torch>=2.0.0",
    "python-chess>=1.9.0",
    "pyyaml>=6.0",
    "h5py>=3.0.0",
    "tqdm>=4.60.0",
    "tensorboard>=2.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.pytest.ini_options]
testpaths = ["training/tests", "selfplay/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
